Move ProductScopes into a Concern
This fixes issues with Rails 6 and Zeitwerk trying to autoload Product::Scopes class. Also is much cleaner and more rails way
