+++ core/app/models/spree/product.rb
@@ -22,6 +22,8 @@ module Spree
   class Product < Spree::Base
     extend FriendlyId
     include ActsAsTaggable
+    include Spree::ProductScopes
+
     friendly_id :slug_candidates, use: :history
 
     acts_as_paranoid
@@ -432,5 +434,3 @@ module Spree
     end
   end
 end
-
-require_dependency 'spree/product/scopes'
